Added more instruction handlers + changed display output
Changed the text output of the display as it will be easier to debug
this way and compare to other emulators. Had to write a function to get
the name of the instruction so that it appears before memory read and
writes, that's about it. Also added more instruction handlers for AND,
ASL, Transfer.